Refactored files into succint folders
