+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./components/styles/index.css";
-import App from "./components/App";
+import App from "./components/App/App";
+import "./assets/styles/index.css";
+
 
 ReactDOM.render(<App />, document.getElementById("root"));
