+++ src/pages/SearchInfoPage/SearchInfo.js
